Solution for XNAT-6012 - Give user ability to see full serialized objects.
Features added to optionally populate delivery history status entries with a serializable payload.
The Logging Action now populates an ACTION_STEP payload with the "filterable event payload", which is the same object acted upon by the subscription JSONPath filter.
If a user wants to see the full serialized object -
	Setup subscription with "Logging Action" selected.
	Trigger corresponding event.
	Review resulting event service history item -> status -> ACTION_STEP
	If the event object can be represented as a string (using ObjectMapper) the payload attribute should display this string.

Relates to XNAT-5216.